+++ test/src/test/java/de/njsm/tinyauth/test/oidc/TokenAsserterWithoutNonce.java
@@ -17,15 +17,7 @@
 
 package de.njsm.tinyauth.test.oidc;
 
-import com.nimbusds.jwt.JWTClaimsSet;
-import de.njsm.tinyauth.test.data.Client;
-import de.njsm.tinyauth.test.data.OidcToken;
-import de.njsm.tinyauth.test.data.User;
 import de.njsm.tinyauth.test.repository.Endpoint;
-
-import java.util.Set;
-
-import static de.njsm.tinyauth.test.oidc.Identifiers.ACCESS_TOKEN;
 
 public class TokenAsserterWithoutNonce implements TokenAsserter {
 
@@ -33,20 +25,6 @@ public class TokenAsserterWithoutNonce implements TokenAsserter {
 
     public TokenAsserterWithoutNonce(Endpoint endpoint) {
         this.endpoint = endpoint;
-    }
-
-    @Override
-    public OidcToken verifyAccessToken(String token, Client client, User user) throws Exception {
-        JWTClaimsSet claims = verifyToken(token);
-        verifyAccessTokenClaims(claims.getClaims(), client, user);
-        return new OidcToken(token, claims);
-    }
-
-    @Override
-    public OidcToken verifyRefreshToken(String token, Client client, User user, Set<String> scopes) throws Exception {
-        JWTClaimsSet claims = verifyRefreshTokenSpecificClaims(token, scopes);
-        verifyAccessTokenClaims(claims.getJSONObjectClaim(ACCESS_TOKEN), client, user);
-        return new OidcToken(token, claims);
     }
 
     @Override