+++ test/src/test/java/de/njsm/tinyauth/test/runtime/SeleniumLifecycleManager.java
@@ -17,8 +17,6 @@
 
 package de.njsm.tinyauth.test.runtime;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.extension.*;
 import org.junit.jupiter.api.extension.support.TypeBasedParameterResolver;
 import org.openqa.selenium.firefox.FirefoxOptions;
@@ -32,9 +30,7 @@ import org.testcontainers.lifecycle.Startables;
 
 import java.io.File;
 
-public class SeleniumLifecycleManager extends TypeBasedParameterResolver<Browser> implements BeforeAllCallback, BeforeEachCallback, ExtensionContext.Store.CloseableResource{
-
-    private static final Logger LOG = LogManager.getLogger(SeleniumLifecycleManager.class);
+public class SeleniumLifecycleManager extends TypeBasedParameterResolver<Browser> implements BeforeAllCallback, BeforeEachCallback, ExtensionContext.Store.CloseableResource {
 
     public static final String REDIRECT_PAGE = "<!doctype html>" +
             "<html>" +
@@ -87,7 +83,7 @@ public class SeleniumLifecycleManager extends TypeBasedParameterResolver<Browser
                 .withCopyToContainer(Transferable.of(REDIRECT_PAGE), "/usr/share/nginx/html/redirect/advanced-client-auth.html")
                 .withNetworkAliases("client")
                 .withNetwork(uut.network())
-                .waitingFor(new HttpWaitStrategy());
+                .waitingFor(new HttpWaitStrategy().forStatusCode(200));
 
         Startables.deepStart(seleniumContainer, nginxContainer).join();
         driver = new RemoteWebDriver(seleniumContainer.getSeleniumAddress(), options);