+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+
+* Scope claims can now go to only some of access token, id token or userinfo
+  ([#76](https://j.njsm.de/git/veenj/tiny-auth/issues/76)).
+
 ## [2.0.0]
 
 ### Added